style: 패키지 명 변경 (step3 -> lotto)
