+++ src/main/java/lotto/domain/LottoCreateStrategy.java
@@ -1,0 +1,17 @@
+package lotto.domain;
+
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+import static lotto.domain.LottoNumber.MAXIMUM_LOTTO_NUMBER;
+import static lotto.domain.LottoNumber.MINIMUM_LOTTO_NUMBER;
+
+public interface LottoCreateStrategy {
+
+    Map<Integer, LottoNumber> lottoNumberMap = IntStream.rangeClosed(MINIMUM_LOTTO_NUMBER, MAXIMUM_LOTTO_NUMBER)
+            .mapToObj(LottoNumber::new)
+            .collect(Collectors.toMap(LottoNumber::lottoNumber, e -> e));
+
+    Lotto createLotto();
+}
