+++ src/main/java/lotto/domain/Lottos.java
@@ -1,0 +1,16 @@
+package lotto.domain;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Lottos {
+    private final List<Lotto> lottos;
+
+    public Lottos(List<Lotto> lottos) {
+        this.lottos = lottos;
+    }
+
+    public List<Lotto> lottos() {
+        return new ArrayList<>(lottos);
+    }
+}
