+++ src/main/java/lotto/ui/InputView.java
@@ -1,0 +1,11 @@
+package lotto.ui;
+
+import lotto.domain.PurchaseAmount;
+import lotto.domain.WinningLotto;
+
+public interface InputView {
+
+    PurchaseAmount readPurchaseAmount();
+
+    WinningLotto readWinningLottoNumbers();
+}
