--- src/main/java/step3/domain/Lottos.java
@@ -1,16 +1,0 @@
-package step3.domain;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Lottos {
-    private final List<Lotto> lottos;
-
-    public Lottos(List<Lotto> lottos) {
-        this.lottos = lottos;
-    }
-
-    public List<Lotto> lottos() {
-        return new ArrayList<>(lottos);
-    }
-}
