--- src/main/java/step3/ui/ResultView.java
@@ -1,16 +1,0 @@
-package step3.ui;
-
-import step3.domain.Lotto;
-import step3.domain.Lottos;
-import step3.domain.PurchaseAmount;
-import step3.domain.WinningLotto;
-
-import java.util.List;
-
-public interface ResultView {
-    void printLottos(List<Lotto> lottos);
-
-    void printLottoTickets(PurchaseAmount purchaseAmount);
-
-    void printWinningResult(WinningResultDTO winningResultDTO);
-}
